ECG scaling and readouts
I have updated so that the heart rates are shown (integer values) from each device as well as either the R wave height above zero or the PPGpeak rate. The PPG rate is divided by 1,000 and in arb. units. It is also negative as increasing blood is a negative signal out of the Verity Sense. I will try and implement a 'short-form' download for those who collect for long periods of time but are only interested in the detected peaks (and their times).